Include return loan feature
The idea is to use deleted logs as a way to mark as blue
the loans that were already returned. This is just a hacky
temporary patch, given that the whole project needs to be rebuilt
if it is to be improved.

Return loan feature works by pasting the datetime of the log on the comments field.
The browser then finds the datetime on the logs table, and retrieves the id of that log
(row's id).

On the server side the api catches the comment message and automatically changes the
colors, deleted tags, and user balance.